+++ tests/testdata/analysis/module_attr_shadowed_param.py
@@ -1,0 +1,6 @@
+# ruff: noqa
+import os
+
+
+def not_a_module_mutation(os):
+    os.getcwd = lambda: "/fake"